refactor: remove double exports
