+++ packages/i18n/src/config/types.ts
@@ -1,4 +1,4 @@
-import { CustomMapping } from 'generaltranslation/types';
+import type { CustomMapping } from 'generaltranslation/types';
 
 /**
  * TODO: this is a react-only type, we need to move this
