+++ packages/i18n/src/index.ts
@@ -1,5 +1,3 @@
 export * from './translation-functions/msg';
 export * from './translation-functions/static';
 export * from './translation-functions/fallbacks';
-export * from './translation-functions/internal';
-export * from './i18n-manager';
