+++ packages/i18n/src/internal-types.ts
@@ -6,8 +6,8 @@ export type {
   StorageAdapter,
 } from './i18n-manager/types';
 
-// Translation functions
-export type * from './translation-functions/types';
+// Translation Options (Function types exported by /types)
+export type * from './translation-functions/types/options';
 
 // Config
 export type * from './config/types';
