+++ packages/i18n/src/internal.ts
@@ -1,2 +1,2 @@
-export * from './translation-functions';
+export * from './translation-functions/internal';
 export * from './i18n-manager';
