--- packages/i18n/src/translation-functions/index.ts
@@ -1,4 +1,0 @@
-export * from './msg';
-export * from './static';
-export * from './fallbacks';
-export * from './internal';
