--- packages/i18n/src/translation-functions/types/index.ts
@@ -1,2 +1,0 @@
-export * from './options';
-export * from './functions';
