+++ packages/node/src/setup/initializeGT.ts
@@ -1,7 +1,7 @@
-import { setI18nManager } from '../i18n-manager/singleton-operations';
+import { setI18nManager } from '../async-i18n-manager/singleton-operations';
 import type { InitializeGTParams } from './types';
-import { AsyncStorageI18nManager } from '../i18n-manager/AsyncStorageI18nManager';
-import { AsyncStorageAdapter } from '../i18n-manager/AsyncStorageAdapter';
+import { AsyncStorageI18nManager } from '../async-i18n-manager/AsyncStorageI18nManager';
+import { AsyncStorageAdapter } from '../async-i18n-manager/AsyncStorageAdapter';
 
 /**
  * Configure GT for node runtime. This must be called to setup GT for node runtime.
