+++ packages/node/src/translation-functions/index.ts
@@ -7,13 +7,13 @@ export { getGT } from 'gt-i18n/internal';
 export { getMessages } from 'gt-i18n/internal';
 
 // Msg Function
-export { msg } from 'gt-i18n/internal';
+export { msg } from 'gt-i18n';
 
 // Static Functions
-export { declareStatic, declareVar } from 'gt-i18n/internal';
+export { declareStatic, declareVar } from 'gt-i18n';
 
 // Decoding Functions
-export { decodeVars, decodeMsg, decodeOptions } from 'gt-i18n/internal';
+export { decodeVars, decodeMsg, decodeOptions } from 'gt-i18n';
 
 // --- TODO: Add the following functions --- //
