+++ packages/node/src/translation-functions/types.ts
@@ -1,13 +1,13 @@
 // --- Function Types --- //
 
 // GT Function
-export type { GTFunctionType } from 'gt-i18n/internal-types';
+export type { GTFunctionType } from 'gt-i18n/types';
 
 // M Function Type
-export type { MFunctionType } from 'gt-i18n/internal-types';
+export type { MFunctionType } from 'gt-i18n/types';
 
 // T Function Type
-export type { TFunctionType } from 'gt-i18n/internal-types';
+export type { TFunctionType } from 'gt-i18n/types';
 
 // Tx Function Type
-export type { TxFunctionType } from 'gt-i18n/internal-types';
+export type { TxFunctionType } from 'gt-i18n/types';
